refactor: extract bootstrap modules from extension.ts
Decomposed the 922-line activate() into focused bootstrap modules:
- bootstrapGenAi: GenAI provider registration
- bootstrapProviders: task provider registration
- bootstrapMobile: mobile server + tunnel setup
- bootstrapSquad: session state, copilot launcher, squad manager
