+++ src/bootstrap/index.ts
@@ -1,0 +1,11 @@
+export { bootstrapGenAi } from './bootstrapGenAi';
+export type { GenAiBootstrapResult } from './bootstrapGenAi';
+
+export { bootstrapMobile } from './bootstrapMobile';
+export type { MobileBootstrapResult } from './bootstrapMobile';
+
+export { bootstrapProviders } from './bootstrapProviders';
+export type { ProvidersBootstrapResult } from './bootstrapProviders';
+
+export { bootstrapSquad } from './bootstrapSquad';
+export type { SquadBootstrapResult } from './bootstrapSquad';
